+++ gui/objects/Movie.cpp
@@ -26,6 +26,7 @@ Movie::Movie(const GuiSettings& settings,
     enabled = settings.get<bool>(GuiSettingsId::IMAGES_SAVE);
     makeAnimation = settings.get<bool>(GuiSettingsId::IMAGES_MAKE_MOVIE);
     outputStep = settings.get<Float>(GuiSettingsId::IMAGES_TIMESTEP);
+    cameraVelocity = settings.get<Vector>(GuiSettingsId::CAMERA_VELOCITY);
     const Path directory(settings.get<std::string>(GuiSettingsId::IMAGES_PATH));
     const Path name(settings.get<std::string>(GuiSettingsId::IMAGES_NAME));
     const Size firstIndex(settings.get<int>(GuiSettingsId::IMAGES_FIRST_INDEX));
@@ -113,6 +114,14 @@ void Movie::save(const Storage& storage, Statistics& stats) {
         const Vector newPos = trackedPos - target + cameraPos;
         params.camera->setPosition(newPos);
         params.camera->setTarget(trackedPos);
+    } else {
+        const Float time = stats.getOr<Float>(StatisticsId::RUN_TIME, 0._f);
+        const Float dt = time - lastFrame;
+        const Vector target = params.camera->getTarget();
+        const Vector cameraPos = params.camera->getFrame().translation();
+        params.camera->setPosition(cameraPos + dt * cameraVelocity);
+        params.camera->setTarget(target + dt * cameraVelocity);
+        lastFrame = time;
     }
 
     const Path path = paths.getNextPath(stats);